requirements.txt
numpy
pandas
matplotlib
scikit-learn
pyyaml
joblib
streamlit
sumolib
traci